Re-design chaincodeType abstraction to use a factory pattern
