+++ openchain/chaincode/platforms/platforms.go
@@ -1,0 +1,24 @@
+package platforms
+
+import (
+	"fmt"
+	"archive/tar"
+	"github.com/openblockchain/obc-peer/openchain/chaincode/platforms/golang"
+	pb "github.com/openblockchain/obc-peer/protos"
+)
+
+type Platform interface {
+	ValidateSpec(spec *pb.ChaincodeSpec) error
+	WritePackage(spec *pb.ChaincodeSpec, tw *tar.Writer) error
+}
+
+func Find(chaincodeType pb.ChaincodeSpec_Type) (Platform, error) {
+
+	switch chaincodeType {
+	case pb.ChaincodeSpec_GOLANG:
+		return &golang.Platform{}, nil
+	default:
+		return nil, fmt.Errorf("Unknown chaincodeType: %s", chaincodeType)
+	}
+
+}
